requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
executing==2.0.0
filelock==3.12.4
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.9.2
gym==0.25.2
gym-notices==0.0.8
hashids==1.3.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.10.0
jedi==0.19.1
jericho @ file:///Data2/rding/jericho
Jinja2==3.1.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mementos==1.3.1
more-itertools==10.1.0
mpmath==1.3.0
msal==1.24.1
multidict==6.0.4
murmurhash==1.0.10
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
openai==0.28.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
preshed==3.0.9
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
TatSu==4.4.0
textworld==1.5.4
thinc==8.2.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.3
yarl==1.9.2
zipp==3.17.0
